pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "debbugs-graphql"
version = "0.1.0"
description = "Interact for debbugs using GraphQL"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.10"
dependencies = [
    "strawberry-graphql[fastapi]~=0.132.1",
    "uvicorn~=0.18.3",
    "requests~=2.28.1",
    "jinja2~=3.1.2",
]

[project.scripts]
debbugs-graphql = "debbugs_graphql.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest~=7.1.3",
    "pytest-icdiff~=0.6",
]

[tool.setuptools]
package-dir = {"" = "src"}
